requirements.txt
pyfunctional
BeautifulSoup4